Anpassungen aus der Live-Session 3

+++ session3/03_file_manage.py
@@ -18,13 +18,14 @@ logdatei = pathlib.Path("rechner_ausgabe.txt")
 if not alte_logs.exists():
     alte_logs.mkdir()
 
-# Logdatei in das Unterverzeichnis kopieren
-shutil.copy2(logdatei, alte_logs)
-# Logdatei im Unterverzeichnis umbenennen mit Zeitstempel
-shutil.move(alte_logs / logdatei,
-            alte_logs / f"rechner_ausgabe_{dt.datetime.now().strftime('%Y%m%d%H%M%S')}.txt")
-# Original Logdatei löschen
-# os.remove(logdatei)
+if logdatei.exists():
+    # Logdatei in das Unterverzeichnis kopieren
+    shutil.copy2(logdatei, alte_logs)
+    # Logdatei im Unterverzeichnis umbenennen mit Zeitstempel
+    shutil.move(alte_logs / logdatei,
+                alte_logs / f"rechner_ausgabe_{dt.datetime.now().strftime('%Y%m%d%H%M%S')}.txt")
+    # Original Logdatei löschen
+    os.remove(logdatei)
 
 # Alte Logs komprimieren
 shutil.make_archive(alte_logs / "logs", "zip", alte_logs)
